feat(ai): integrate AI model configuration and enhance chat functionality in AiService
