+++ src/app/components/dashboard/lawn-health-score/lawn-health-score.component.ts
@@ -1,28 +1,8 @@
-import { Component, computed, inject } from '@angular/core';
-import { rxResource } from '@angular/core/rxjs-interop';
+import { Component, inject } from '@angular/core';
 import { CardModule } from 'primeng/card';
 import { DividerModule } from 'primeng/divider';
 import { DividerDesignTokens } from '@primeng/themes/types/divider';
-import { AnalyticsService } from '../../../services/analytics.service';
-import { EntriesService } from '../../../services/entries.service';
-import { AiService } from '../../../services/ai.service';
-import {
-  calculateLawnHealthScore,
-  isCurrentlyGrowingSeason
-} from '../../../utils/lawnHealthUtils';
-import { getPoundsOfNInFertilizerApp } from '../../../utils/lawnCalculatorUtils';
-import {
-  LawnHealthScoreFactors,
-  MonthlyData
-} from '../../../types/lawnHealthScore.types';
-import {
-  subMonths,
-  getMonth,
-  getYear,
-  isSameMonth,
-  isSameYear,
-  differenceInDays
-} from 'date-fns';
+import { LawnHealthScoreService } from '../../../services/lawn-health-score.service';
 
 @Component({
   selector: 'lawn-health-score',
@@ -31,203 +11,15 @@ import {
   styleUrl: './lawn-health-score.component.scss'
 })
 export class LawnHealthScoreComponent {
-  private _analyticsService = inject(AnalyticsService);
-  private _entriesService = inject(EntriesService);
-  private _aiService = inject(AiService);
+  private _lawnHealthScoreService = inject(LawnHealthScoreService);
 
-  public analyticsData = this._analyticsService.analyticsData;
-  public lastMowDate = this._entriesService.lastMow;
-  public lastProductApp = this._entriesService.lastProductApp;
-
-  public healthScoreFactors = computed((): LawnHealthScoreFactors => {
-    const lastMow = this.lastMowDate.value();
-    const lastProductAppData = this.lastProductApp.value();
-
-    const currentMonth = getMonth(new Date()) + 1;
-    const isGrowingSeason = isCurrentlyGrowingSeason();
-
-    const daysSinceLastMow = lastMow?.lastMowDate
-      ? differenceInDays(new Date(), lastMow.lastMowDate.toString())
-      : 999;
-
-    const daysSinceLastFertilizer = lastProductAppData?.lastProductAppDate
-      ? differenceInDays(
-          new Date(),
-          lastProductAppData.lastProductAppDate.toString()
-        )
-      : 999;
-
-    const monthlyData = this.getLast3MonthsData();
-
-    return {
-      daysSinceLastMow,
-      daysSinceLastFertilizer,
-      monthlyData,
-      currentMonth,
-      isGrowingSeason
-    };
-  });
-
-  public lawnHealthScore = computed(() => {
-    const analyticsData = this.analyticsData.value();
-    const factors = this.healthScoreFactors();
-
-    return calculateLawnHealthScore(analyticsData, factors);
-  });
-
-  public aiDescriptionResource = rxResource({
-    params: () => {
-      const scoreData = this.lawnHealthScore();
-      return scoreData.totalScore > 0
-        ? {
-            score: scoreData.totalScore,
-            grade: scoreData.grade
-          }
-        : undefined;
-    },
-    stream: ({ params }: { params: { score: number; grade: string } }) => {
-      const currentDate = new Date().toLocaleDateString('en-US', {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric'
-      });
-
-      const prompt = `Today is ${currentDate}. Generate a brief, encouraging lawn care description (max 30 words) for someone with a lawn health score of ${params.score}/100 (Grade: ${params.grade}). Focus on positive reinforcement and seasonal next steps.`;
-      return this._aiService.sendChatMessage(prompt);
-    }
-  });
-
-  public finalDescription = computed(() => {
-    const aiResource = this.aiDescriptionResource;
-    const fallbackDesc = this.lawnHealthScore().description;
-
-    if (aiResource.hasValue() && aiResource.value()) {
-      return aiResource.value();
-    }
-
-    return fallbackDesc;
-  });
-
-  public scoreColor = computed(() => {
-    const score = this.lawnHealthScore().totalScore;
-    if (score >= 90) return 'var(--p-green-500)';
-    if (score >= 80) return 'var(--p-blue-500)';
-    if (score >= 70) return 'var(--p-yellow-500)';
-    if (score >= 60) return 'var(--p-orange-500)';
-    return 'var(--p-red-500)';
-  });
+  public lawnHealthScore = this._lawnHealthScoreService.lawnHealthScore;
+  public finalDescription = this._lawnHealthScoreService.finalDescription;
+  public scoreColor = this._lawnHealthScoreService.scoreColor;
 
   public dividerDt: DividerDesignTokens = {
     horizontal: {
       margin: 'none'
     }
   };
-
-  private getLast3MonthsData(): [MonthlyData, MonthlyData, MonthlyData] {
-    const now = new Date();
-
-    const months = [];
-    for (let i = 0; i < 3; i++) {
-      const targetDate = subMonths(now, i);
-      const month = getMonth(targetDate) + 1;
-      const year = getYear(targetDate);
-
-      months.push({
-        mowingFrequency: this.getMowingFrequencyForMonth(month, year),
-        nitrogenAmount: this.getNitrogenForMonth(month, year),
-        fertilizerApplications: this.getFertilizerApplicationsForMonth(
-          month,
-          year
-        ),
-        entriesCount: this.getEntriesForMonth(month, year),
-        month,
-        year,
-        isGrowingSeason: month >= 4 && month <= 10
-      });
-    }
-
-    return months as [MonthlyData, MonthlyData, MonthlyData];
-  }
-
-  private getMowingFrequencyForMonth(month: number, year: number): number {
-    const analyticsData = this.analyticsData.value();
-    if (!analyticsData?.mowingAnalyticsData) return 0;
-
-    const monthData = analyticsData.mowingAnalyticsData.find(
-      (item) => item.month === month && item.year === year
-    );
-
-    return monthData ? +monthData.mowCount : 0;
-  }
-
-  private getNitrogenForMonth(month: number, year: number): number {
-    const analyticsData = this.analyticsData.value();
-    if (!analyticsData?.fertilizerTimelineData) return 0;
-
-    return analyticsData.fertilizerTimelineData
-      .filter((app) => {
-        const appDate = new Date(app.applicationDate);
-        const targetDate = new Date(year, month - 1, 1);
-        return (
-          isSameMonth(appDate, targetDate) && isSameYear(appDate, targetDate)
-        );
-      })
-      .reduce((total, app) => {
-        const isLiquid =
-          app.productQuantityUnit === 'oz' ||
-          app.productQuantityUnit === 'fl oz';
-        const productWeight = isLiquid
-          ? app.productQuantity * 0.1
-          : app.productQuantity;
-
-        const nitrogenValue = getPoundsOfNInFertilizerApp({
-          poundsOfProduct: productWeight,
-          guaranteedAnalysisOfProduct: app.guaranteedAnalysis,
-          totalSquareFeet: app.totalSquareFeet
-        });
-
-        return total + nitrogenValue;
-      }, 0);
-  }
-
-  private getFertilizerApplicationsForMonth(
-    month: number,
-    year: number
-  ): number {
-    const analyticsData = this.analyticsData.value();
-    if (!analyticsData?.fertilizerTimelineData) return 0;
-
-    return analyticsData.fertilizerTimelineData.filter((app) => {
-      const appDate = new Date(app.applicationDate);
-      const targetDate = new Date(year, month - 1, 1);
-      return (
-        isSameMonth(appDate, targetDate) && isSameYear(appDate, targetDate)
-      );
-    }).length;
-  }
-
-  private getEntriesForMonth(month: number, year: number): number {
-    const analyticsData = this.analyticsData.value();
-    if (
-      !analyticsData?.mowingAnalyticsData &&
-      !analyticsData?.fertilizerTimelineData
-    )
-      return 2;
-
-    const mowingEntries =
-      analyticsData?.mowingAnalyticsData?.find(
-        (item) => item.month === month && item.year === year
-      )?.mowCount || '0';
-
-    const fertilizerEntries =
-      analyticsData?.fertilizerTimelineData?.filter((app) => {
-        const appDate = new Date(app.applicationDate);
-        const targetDate = new Date(year, month - 1, 1);
-        return (
-          isSameMonth(appDate, targetDate) && isSameYear(appDate, targetDate)
-        );
-      }).length || 0;
-
-    return +mowingEntries + fertilizerEntries;
-  }
 }
